Rename things to not surface TRPC for now
