+++ src/lib/client/createClient.ts
@@ -12,3 +12,4 @@ export const trpc = createTRPCProxyClient<AppRouter>({
     }),
   ],
 });
+export const client = trpc;
