+++ src/pages/api/trpc/[trpc].ts
@@ -1,6 +1,6 @@
 import type { APIRoute } from 'astro';
-import { createAstroTRPCApiHandler } from '../../../lib/trpc';
-import { appRouter } from '../../../lib/trpc/appRouter';
+import { createAstroTRPCApiHandler } from '../../../lib/client';
+import { appRouter } from '../../../lib/client/appRouter';
 
 // Return 'all' as an Astro api route
 export const all = createAstroTRPCApiHandler({
